requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aniso8601==10.0.1
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-RESTful==0.3.10
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
grpcio==1.71.0
grpcio-status==1.71.0
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
pyparsing==3.2.3
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
rfc3986==1.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
